// angular-sample/angular-webapp/client/app/risk-assessment/models/risk-assessment-substance.ts
export class RASubstance {
    Id: string;
    CompanyId: string;
    Description: string;
    Substance: string;
    Manufacturer: string;
    UsedFor: string;
    ExposureLimits: string;
    ReferenceNumber: string;
    Quantity: string;
    PeopleAffected: string;
    AttachmentId: string;
    RiskAssessmentId: string;
    IsDeleted: boolean;
    Mig_EmergencyContactNumber: number;
}