// angular-sample/angular-webapp/client/app/risk-assessment/models/risk-assessment-coshh-inventory.ts
export class RACoshhInventory {
    Id: string;
    CompanyId: string;
    Description: string;
    Substance: string;
    Manufacturer: string;
    UsedFor: string;
    ExposureLimits: string;
    ReferenceNumber: string;
    Quantity: string;
    PeopleAffected: string;
    AttachmentId: string;
    RiskAssessmentId: string;
    Author: string;
    IsDeleted:boolean;
    IsExample:boolean;
    Mig_EmergencyContactNumber: number;
}